finalize api and write README.md
