+++ planning/autoware_avoidance_target_detector/package.xml
@@ -21,13 +21,14 @@
   <depend>autoware_trajectory</depend>
   <depend>autoware_utils</depend>
   <depend>autoware_utils_geometry</depend>
-  <depend>autoware_vehicle_info_utils</depend>
   <depend>lanelet2_core</depend>
+  <depend>lanelet2_io</depend>
   <depend>lanelet2_routing</depend>
   <depend>lanelet2_traffic_rules</depend>
-  <depend>libboost-dev</depend>
   <depend>rclcpp</depend>
   <depend>rclcpp_components</depend>
+
+  <exec_depend>j6_gen2_description</exec_depend>
 
   <export>
     <build_type>ament_cmake</build_type>
